app/requirements.txt
streamlit
pandas
python-dotenv
requests
asyncio
aiohttp